refactor: split oversized controllers
